test for nothing to happen

+++ cgp3-prep/cgp3-prep/tesselate/ffd.h
@@ -42,7 +42,7 @@ private:
 
 public:
 
-    uint nChoosek( unsigned n, unsigned k );
+    float nChoosek( float n, float k );
 
     ShapeGeometry geom;         ///< renderable version of non-active lattice
     ShapeGeometry activegeom;   ///< renderable version of active lattice (selected control points)
